requirements.txt
aiogram=2.13
Pillow=8.2.0
torch=1.9.0+cpu
torchvision=0.10.0+cpu
